% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_samples.R
\name{collapse_samples}
\alias{collapse_samples}
\title{Combines samples based on a column in the colData}
\usage{
collapse_samples(m, col = NULL)
}
\arguments{
\item{m}{\code{\link{methrix}} object}

\item{col}{a column in the colData, with new sample names. Samples with the same names will be merged and those with NA will be removed}
}
\value{
a \code{\link{methrix}} object
}
\description{
Combines samples based on a column in the colData
}
\details{
Takes \code{\link{methrix}} object and a column name and combines samples with the same name in the annotataion.
}
\examples{
data("methrix_data")
collapse_samples(m = methrix_data, col="Condition")
}
