% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{combine_methrix}
\alias{combine_methrix}
\title{Combine methrix objects}
\usage{
combine_methrix(m1, m2, by = c("row", "col"))
}
\arguments{
\item{m1}{Frist \code{\link{methrix}} object}

\item{m2}{Second \code{\link{methrix}} object}

\item{by}{The direction of combine. 'column' (cbind) combines samples with same regions, 'row' combines different regions,
e.g. different chromosomes.}
}
\value{
An object of class \code{\link{methrix}}
}
\description{
Combine methrix objects
}
\details{
Takes two \code{\link{methrix}} objects and combines them row- or column-wise
}
