% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{convert_methrix}
\alias{convert_methrix}
\title{Converts an in-memory object to an on-disk HDF5 object.}
\usage{
convert_methrix(m = NULL)
}
\arguments{
\item{m}{An object of class \code{\link{methrix}}}
}
\value{
An object of class \code{\link{methrix}}, HDF5 format
}
\description{
Converts an in-memory object to an on-disk HDF5 object.
}
\details{
Takes a \code{\link{methrix}} object and returns with the same object with delayed array assay slots
with HDF5 backend. Might take long time!
}
\examples{
data(methrix_data)
m2 <- convert_methrix(m=methrix_data)
}
