% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{filter_methrix}
\alias{filter_methrix}
\title{Filter matrices by coverage}
\usage{
filter_methrix(m, cov_thr = 1, min_samples = 1, n_threads = 4)
}
\arguments{
\item{m}{\code{\link{methrix}} object}

\item{cov_thr}{minimum coverage required to call a loci covered}

\item{min_samples}{At-least these many samples should have a loci with coverage >= \code{cov_thr}}

\item{n_threads}{number of threads to use. Default 4.}
}
\description{
Filter matrices by coverage
}
\details{
Takes \code{\link{methrix}} object and filters CpGs based on coverage statistics
}
