% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{get_chr_summary}
\alias{get_chr_summary}
\title{Extract CpGs covered per chromosome information}
\usage{
get_chr_summary(m = NULL)
}
\arguments{
\item{m}{\code{\link{methrix}} object}
}
\description{
Extract CpGs covered per chromosome information
}
\details{
Takes \code{\link{methrix}} object and returns a table of CpGs covered per chromosome
}
\examples{
data("methrix_data")
get_chr_summary(methrix_data)
}
