% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{get_region_summary}
\alias{get_region_summary}
\title{Extracts methylation of regions of interest}
\usage{
get_region_summary(m, regions = NULL, type = NULL, how = NULL)
}
\arguments{
\item{regions}{genomic regions to be summarized. Could be a data.table with 3 columns (chr, start, end) or a \code{\link{GRanges}} object}

\item{type}{data that should be summarized. Coule be methylation or coverage.}

\item{how}{mathematical function by which regions should be summarized. Can be one of the following: mean, sum, max, min}
}
\description{
Extracts methylation of regions of interest
}
\details{
Takes \code{\link{methrix}} object and summarizes regions
}
