% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{get_stat}
\alias{get_stat}
\title{Estimate descriptive statistics}
\usage{
get_stat(m, per_chr = FALSE)
}
\arguments{
\item{m}{\code{\link{methrix}} object}

\item{per_chr}{Estimate stats per chromosome. Default FALSE}
}
\description{
Estimate descriptive statistics
}
\details{
Calculate descriptive statistics
}
\seealso{
\code{\link{plot_stats}}
}
