% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modkit_v2.R
\name{get_tabix_chromosomes}
\alias{get_tabix_chromosomes}
\title{Get list of chromosomes present in tabix-indexed files}
\usage{
get_tabix_chromosomes(files, verbose = FALSE)
}
\arguments{
\item{files}{Character vector of file paths}

\item{verbose}{Print messages}
}
\value{
Character vector of chromosome names
}
\description{
Get list of chromosomes present in tabix-indexed files
}
\keyword{internal}
