% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_object.R
\docType{methods}
\name{getChrSummary}
\alias{getChrSummary}
\alias{getChrSummary,methrix-method}
\title{extract CpGs per chromosome stats}
\usage{
getChrSummary(x)

\S4method{getChrSummary}{methrix}(x)
}
\arguments{
\item{x}{An object of class methrix}
}
\value{
CpG per chromosome table
}
\description{
extract CpGs per chromosome stats
}
