% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_methods.R
\docType{methods}
\name{getCoverageMatrix}
\alias{getCoverageMatrix}
\alias{getCoverageMatrix,methrix-method}
\title{extract methylation matrix}
\usage{
getCoverageMatrix(x)

\S4method{getCoverageMatrix}{methrix}(x)
}
\arguments{
\item{x}{An object of class methrix}
}
\value{
coverage methrix
}
\description{
extract methylation matrix
}
