% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_methods.R
\docType{methods}
\name{getCpgLoci}
\alias{getCpgLoci}
\alias{getCpgLoci,methrix-method}
\title{extract CpG Loci}
\usage{
getCpgLoci(x)

\S4method{getCpgLoci}{methrix}(x)
}
\arguments{
\item{x}{An object of class methrix}
}
\value{
An object of class DataFrame
}
\description{
extract CpG Loci
}
