% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_methods.R
\docType{methods}
\name{getMethylationMatrix}
\alias{getMethylationMatrix}
\alias{getMethylationMatrix,methrix-method}
\title{extract methylation matrix}
\usage{
getMethylationMatrix(x)

\S4method{getMethylationMatrix}{methrix}(x)
}
\arguments{
\item{x}{An object of class methrix}
}
\value{
methylation methrix
}
\description{
extract methylation matrix
}
