% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_object.R
\docType{class}
\name{methrix-class}
\alias{methrix-class}
\alias{methrix}
\title{Class methrix}
\description{
S4 class Methrix
}
\section{Slots}{

\describe{
\item{\code{beta}}{beta matrix}

\item{\code{cov}}{coverage matrix}

\item{\code{metadata}}{genome: the name of the BSgenome that was used to extract CpGs, isHDF5: is it stored in HDF5 Array format}

\item{\code{sample_annotation}}{sample annotation data frame}

\item{\code{position}}{chromosome, start and strand information in a data.table format}
}}

