% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix-package.R
\docType{package}
\name{methrix-package}
\alias{methrix-package}
\title{methrix: Analysis of DNA methylation data}
\description{
methrix provides functionality for reading, processing, and analyzing
DNA methylation data from various sources including ModKit bedMethyl files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CompEpigen/methrix}
  \item Report bugs at \url{https://github.com/CompEpigen/methrix/issues}
}

}
\author{
\strong{Maintainer}: Anand Mayakonda \email{anand_mt@hotmail.com} (\href{https://orcid.org/0000-0003-1162-687X}{ORCID})

Authors:
\itemize{
  \item Reka Toth \email{r.toth@dkfz.de} (\href{https://orcid.org/0000-0002-6096-1052}{ORCID})
}

Other contributors:
\itemize{
  \item Rajbir Batra [contributor]
  \item Clarissa Feuerstein-Akgöz [contributor]
  \item Joschka Hey [contributor]
  \item Maximilian Schönung [contributor]
  \item Pavlo Lutsik [contributor]
}

}
\keyword{internal}
