% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix-plot.R
\name{methrix_coverage}
\alias{methrix_coverage}
\title{Covergae QC Plots}
\usage{
methrix_coverage(m, type = c("hist", "dens"), pheno = NULL,
  perSample = FALSE, lim = 100)
}
\arguments{
\item{m}{Input \code{\link{methrix}} object}

\item{type}{Choose between "hist" (histogram) or "dens" (density plot).}

\item{pheno}{Column name of colData(m). Will be used as a factor to color different groups in the violin plot.}

\item{perSample}{Color the plots in a sample-wise manner?}

\item{lim}{Maximum coverage value to be plotted.}
}
\value{
ggplot2 object
}
\description{
Covergae QC Plots
}
\examples{
data("mm9_bsmap")
methrix_coverage(m = mm9_bsmap)
}
