% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_data.R
\docType{data}
\name{methrix_data_h5}
\alias{methrix_data_h5}
\title{WGBS for colon cancer, chr21 and chr22, HDF5 format}
\format{An object of class \code{"methrix"}}
\usage{
data("methrix_data_h5")
}
\description{
This is a subset of original `bsseqData` converted to `methrix` containing Whole-genome bisulfite sequencing data (WGBS) for colon cancer on chromosome 21 and 22.
}
\examples{
data("methrix_data_h5")
methrix_data_h5
}
\references{
Hansen, K. D. et al. (2011) Increased methylation variation in epigenetic domains across cancer types. Nature Genetics 43, 768-775.
}
\keyword{datasets,}
\keyword{h5}
