% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix-plot.R
\name{methrix_density}
\alias{methrix_density}
\title{Density Plot of \eqn{\beta}-Values}
\usage{
methrix_density(m, ranges = NULL, pheno = NULL, return_ggplot = F,
  bw.adjust = 2)
}
\arguments{
\item{m}{Input \code{\link{methrix}} object}

\item{ranges}{\code{\link{GRanges}} object with GenomicRegions which should be used for plotting.}

\item{pheno}{Column name of colData(m). Will be used as a factor to color different groups in the violin plot.}

\item{return_ggplot}{Return a \code{\link{ggplot2}} object.}

\item{bw.adjust}{Multiplicate bandwide adjustment. See \code{\link{geom_density}} for more information}
}
\description{
Density Plot of \eqn{\beta}-Values
}
