% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mehtrix-differential.R
\name{methrix_differential}
\alias{methrix_differential}
\title{Differential Methylation Calling with \code{\link{DSS}}}
\usage{
methrix_differential(m, pheno = NULL, smooth = FALSE, s_span = 500,
  delta = 0.1, p = 0.05, minlen = 50, minCG = 3, mergeDis = 500,
  pct.sig = 0.5, parallel = FALSE, chrWise = FALSE, ncore = 1)
}
\arguments{
\item{m}{Input \code{\link{methrix}} object.}

\item{pheno}{Column name of colData(m). The values of this column will be used as classes to call differential methylation.}

\item{smooth}{Should the methylation values be smoothed?}

\item{s_span}{Smoothing span. See \code{\link{DMLtest}}. Default 500.}

\item{delta}{Methylation difference to call a DML/DMR significant. Default 0.1.}

\item{p}{P-value cut-off. Default 0.05.}

\item{minlen}{Minimal length of a DMR. Default 50 bp.}

\item{minCG}{Minimal number of CpGs in a DMR. Default 3.}

\item{mergeDis}{Maximal distance of two consecutive DMRs to be merged. Default 500 bp.}

\item{pct.sig}{Minimum percentage of significant CpGs in a DMR. Default 0.5.}

\item{parallel}{Execute job parallel? Just possible if chrWise = TRUE. Default FALSE.}

\item{chrWise}{Compute job chromosome-wise and compute FDR adjusted p-values afterwards? Default FALSE.}

\item{ncore}{Number of cores if executed parallel. Default 1.}
}
\value{
list of DMRs
}
\description{
Differential Methylation Calling with \code{\link{DSS}}
}
\details{
Wrapper function for differential methylation calling using \code{\link{DSS}}
}
