% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_report.R
\name{methrix_report}
\alias{methrix_report}
\title{Creates a detailed interative html summary report from Methrix object}
\usage{
methrix_report(meth, output_dir = NULL, plot_beta_dist = FALSE,
  n_thr = 4)
}
\arguments{
\item{meth}{\code{\link{methrix}} object}

\item{output_dir}{Output directory name where the files should be saved. If \code{NULL} creats a \code{tempdir}}

\item{plot_beta_dist}{Default FALSE. This is a time consuming and writes huge density files required for plotting.}

\item{n_thr}{Default 4. Only used if \code{plot_beta_dist} is TRUE}
}
\description{
Creates a detailed interative html summary report from Methrix object.
If the directory contains required files (from previous run), it directly proceeds to generate html report.
}
