% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modkit_v2.R
\name{parse_bed_file}
\alias{parse_bed_file}
\title{Parse BED file into data.frame}
\usage{
parse_bed_file(bed_path)
}
\arguments{
\item{bed_path}{Path to BED file}
}
\value{
data.frame with chr, start, end columns (0-based)
}
\description{
Parse BED file into data.frame
}
\keyword{internal}
