% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_plot.R
\name{plot_stats}
\alias{plot_stats}
\title{Plot descriptive statistics}
\usage{
plot_stats(
  plot_dat,
  what = "M",
  stat = "mean",
  ignore_chr = NULL,
  samples = NULL,
  n_col = NULL,
  n_row = NULL
)
}
\arguments{
\item{plot_dat}{results from \code{\link{get_stats}}}

\item{what}{Can be \code{M} or \code{C}. Default \code{M}}

\item{stat}{Can be \code{mean} or \code{median}. Default \code{mean}}

\item{ignore_chr}{Chromsomes to ignore. Default \code{NULL}}

\item{samples}{Use only these samples. Default \code{NULL}}

\item{n_col}{number of columns. Passed to `facet_wrap`}

\item{n_row}{number of rows. Passed to `facet_wrap`}
}
\value{
ggplot2 object
}
\description{
Plot descriptive statistics
}
\details{
plot descriptive statistics results from \code{\link{get_stats}}
}
\examples{
data('methrix_data')
gs = get_stats(methrix_data)
plot_stats(gs)
}
\seealso{
\code{\link{get_stats}}
}
