% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bedgraphs.R
\name{read_bedgraphs}
\alias{read_bedgraphs}
\title{Versatile BedGraph reader.}
\usage{
read_bedgraphs(files = NULL, pipeline = NULL, chr_idx = NULL,
  start_idx = NULL, end_idx = NULL, beta_idx = NULL, M_idx = NULL,
  U_idx = NULL, strand_idx = NULL, cov_idx = NULL, coldata = NULL,
  h5 = FALSE, h5_dir = NULL)
}
\arguments{
\item{files}{bedgraphs or coverage files generated by Bismark. Default NULL}

\item{pipeline}{Default NULL. Can be "Bismark" or "MethylDeckal".}

\item{h5}{Should the coverage and methylation matrices be stored as "HDF5Array"}

\item{h5_dir}{directory to store H5 based object}

\item{loci}{NULL (Default) or a file with  "chromsome" and "start" position. If provided matrix will be generated for these CpGs}

\item{colData}{An optional DataFrame describing the samples. Row names, if present, become the column names of the matrix. If NULL, then a DataFrame will be created with basename of files used as the row names.}

\item{file_format}{are input files bedgraph or coverage files. Default 'auto' predicts.}

\item{vectorize}{Default TRUE. If FALSE uses for loop (much slower but consumes lesser memory)}

\item{n_cores}{number of cores to use. Default NULL. Uses available cores - 1}

\item{shift_start}{Default FALSE. If coordinates are one-based, set this is to TRUE, and start positions will be shifted to its previous base (0-based).}
}
\description{
Versatile BedGraph reader.
}
\details{
Reads bismark coverage files and generates methylation and coverage matrices. Optionally arrays can be serialized as on-disk HDFS5 arrays.
}
