% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_methrix.R
\name{read_bismark}
\alias{read_bismark}
\title{Generates methrix from bismark coverage files}
\usage{
read_bismark(files = NULL, colData = NULL, file_format = "auto",
  loci = NULL, vectorize = TRUE, n_cores = NULL,
  shift_start = FALSE, h5 = FALSE, h5_dir = NULL)
}
\arguments{
\item{files}{bedgraphs or coverage files generated by Bismark. Default NULL}

\item{colData}{An optional DataFrame describing the samples. Row names, if present, become the column names of the matrix. If NULL, then a DataFrame will be created with basename of files used as the row names.}

\item{file_format}{are input files bedgraph or coverage files. Default 'auto' predicts.}

\item{loci}{NULL (Default) or a file with  "chromsome" and "start" position. If provided matrix will be generated for these CpGs}

\item{vectorize}{Default TRUE. If FALSE uses for loop (much slower but consumes lesser memory)}

\item{n_cores}{number of cores to use. Default NULL. Uses available cores - 1}

\item{shift_start}{Default FALSE. If coordinates are one-based, set this is to TRUE, and start positions will be shifted to its previous base (0-based).}

\item{h5}{Should the coverage and methylation matrices be stored as "HDF5Array"}

\item{h5_dir}{directory to store H5 based object}
}
\description{
Generates methrix from bismark coverage files
}
\details{
Reads bismark coverage files and generates methylation and coverage matrices. Optionally arrays can be serialized as on-disk HDFS5 arrays.
}
