% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{subset_methrix}
\alias{subset_methrix}
\title{Subsets \code{\link{methrix}} object based on given conditions.}
\usage{
subset_methrix(m, regions = NULL, contigs = NULL, samples = NULL)
}
\arguments{
\item{m}{\code{\link{methrix}} object}

\item{regions}{genomic regions to subset by. Could be a data.table with 3 columns (chr, start, end) or a \code{\link{GRanges}} object}

\item{contigs}{chromosome names to subset by}

\item{samples}{sample names to subset by}
}
\description{
Subsets \code{\link{methrix}} object based on given conditions.
}
\details{
Takes \code{\link{methrix}} object and filters CpGs based on coverage statistics
}
