% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modkit_v2.R
\name{validate_regions_chromosomes}
\alias{validate_regions_chromosomes}
\title{Validate that regions reference chromosomes present in input files}
\usage{
validate_regions_chromosomes(
  intervals_df,
  files,
  chrom_sizes_path,
  verbose = FALSE
)
}
\arguments{
\item{intervals_df}{data.frame with chr, start, end}

\item{files}{Input bedMethyl files}

\item{chrom_sizes_path}{Path to chromosome sizes file}

\item{verbose}{Verbose output}
}
\value{
Filtered intervals_df with only valid chromosomes
}
\description{
Validate that regions reference chromosomes present in input files
}
\keyword{internal}
