% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bedgraphs.R
\name{write_bedgraphs}
\alias{write_bedgraphs}
\title{Creates a detailed interative html summary report from Methrix object}
\usage{
write_bedgraphs(m)
}
\arguments{
\item{meth}{\code{\link{methrix}} object}

\item{output_dir}{Output directory name where the files should be saved. If \code{NULL} creats a \code{tempdir}}

\item{n_thr}{Default 4. Only used if \code{plot_beta_dist} is TRUE}
}
\description{
Creates a detailed interative html summary report from Methrix object
}
