% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bigwigs.R
\name{write_bigwigs}
\alias{write_bigwigs}
\title{Exports methrix object as bigWigs}
\usage{
write_bigwigs(m, output_dir = getwd(), samp_names = NULL)
}
\arguments{
\item{m}{\code{\link{methrix}} object}

\item{output_dir}{Output directory name where the files should be saved. Default getwd()}

\item{samp_names}{sample names to export}
}
\description{
Exports methrix object as bigWigs
}
\examples{
data('methrix_data')
write_bigwigs(m = methrix_data, output_dir = './temp')
}
