% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{coverage_filter}
\alias{coverage_filter}
\title{Filter matrices by coverage}
\usage{
coverage_filter(m, cov_thr = 1, min_samples = 1, group = NULL)
}
\arguments{
\item{m}{\code{\link{methrix}} object}

\item{cov_thr}{minimum coverage required to call a loci covered}

\item{min_samples}{At-least these many samples should have a loci with coverage >= \code{cov_thr}}

\item{group}{a column name from sample annotation that defines groups. In this case, the number of min_samples will be 
tested group-wise.}
}
\value{
An object of class \code{\link{methrix}}
}
\description{
Filter matrices by coverage
}
\details{
Takes \code{\link{methrix}} object and filters CpGs based on coverage statistics
}
\examples{
data('methrix_data')
#keep only CpGs which are covered by at-least 1 read across 3 samples
coverage_filter(m = methrix_data, cov_thr = 1, min_samples = 3)
}
