% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CpGs.R
\name{extract_CPGs}
\alias{extract_CPGs}
\title{Extracts all CpGs from a genome}
\usage{
extract_CPGs(ref_genome = NULL)
}
\arguments{
\item{ref_genome}{BSgenome object or name of the installed BSgenome package. Example: BSgenome.Hsapiens.UCSC.hg19}
}
\value{
a list of data.table containing number of CpG's and contig lengths
}
\description{
Extracts all CpGs from a genome
}
\examples{
\dontrun{
hg19_cpgs = methrix::extract_CPGs(ref_genome = 'BSgenome.Hsapiens.UCSC.hg19')
}
}
