% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{get_region_summary}
\alias{get_region_summary}
\title{Extract and summarize methylation or coverage info by regions of interest}
\usage{
get_region_summary(
  m,
  regions = NULL,
  type = "M",
  how = "mean",
  overlap_type = "within",
  na_rm = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{m}{\code{\link{methrix}} object}

\item{regions}{genomic regions to be summarized. Could be a data.table with 3 columns (chr, start, end) or a \code{GenomicRanges} object}

\item{type}{matrix which needs to be summarized. Coule be `M`, `C`. Default 'M'}

\item{how}{mathematical function by which regions should be summarized. Can be one of the following: mean, sum, max, min. Default 'mean'}

\item{overlap_type}{defines the type of the overlap of the CpG sites with the target region. Default value is `within`. For detailed description,
see the \code{foverlaps} function of the \code{\link{data.table}} package.}

\item{na_rm}{Remove NA's? Default \code{TRUE}}

\item{verbose}{Default TRUE}
}
\value{
a coverage or methylation matrix
}
\description{
Extract and summarize methylation or coverage info by regions of interest
}
\details{
Takes \code{\link{methrix}} object and summarizes regions
}
\examples{
data('methrix_data')
get_region_summary(m = methrix_data,
regions = data.table(chr = 'chr21', start = 27867971, end =  27868103),
type = 'M', how = 'mean')
}
