% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{get_stats}
\alias{get_stats}
\title{Estimate descriptive statistics}
\usage{
get_stats(m, per_chr = TRUE)
}
\arguments{
\item{m}{\code{\link{methrix}} object}

\item{per_chr}{Estimate stats per chromosome. Default TRUE}
}
\value{
data.table of summary stats
}
\description{
Estimate descriptive statistics
}
\details{
Calculate descriptive statistics
}
\examples{
data('methrix_data')
get_stats(methrix_data)
}
\seealso{
\code{\link{plot_stats}}
}
