% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_object.R
\docType{class}
\name{methrix-class}
\alias{methrix-class}
\alias{methrix}
\title{Class methrix}
\description{
S4 class Methrix
}
\section{Slots}{

\describe{
\item{\code{SE}}{summarized experiment object}

\item{\code{CpGs}}{number of CpGs}

\item{\code{samples}}{number of samples}

\item{\code{h5}}{are matrices are HDF5Array}
}}

