% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_object.R
\docType{class}
\name{methrix-class}
\alias{methrix-class}
\alias{methrix}
\title{Class methrix}
\description{
S4 class Methrix
}
\section{Slots}{

\describe{
\item{\code{assays}}{A list of tw omatrices containing "Methylation" and "Coverage" information}

\item{\code{elementMetadata}}{A DataFrame describing rows in correspoding assay matrices.}

\item{\code{colData}}{genome: the name of the BSgenome that was used to extract CpGs, isHDF5: is it stored in HDF5 Array format}

\item{\code{metadata}}{a list of meta data associated with the assays}

\item{\code{NAMES}}{NULL}
}}

