% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix-package.R
\docType{package}
\name{methrix-package}
\alias{methrix}
\alias{methrix-package}
\title{methrix: Fast and efficient summarization of generic bedGraph files from Bisufite sequencing}
\description{
Bedgraph files generated by Bisulfite pipelines often come in various flavors. Critical downstream step requires summarization of these files into methylation/coverage matrices This step of data aggregation is done by Methrix, including many other useful downstream functions.
}
\author{
\strong{Maintainer}: Anand Mayakonda \email{a.mayakonda@dkfz.de}

Authors:
\itemize{
  \item Reka Toth \email{r.toth@dkfz.de}
}

Other contributors:
\itemize{
  \item Pavlo Lutsik [contributor]
  \item Joschka Hey [contributor]
  \item Maximilian Schönung [contributor]
}

}
\keyword{internal}
