% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{methrix2bsseq}
\alias{methrix2bsseq}
\title{Convert \code{\link{methrix}} to \code{\link{bsseq}} object}
\usage{
methrix2bsseq(m)
}
\arguments{
\item{m}{\code{\link{methrix}} object}
}
\value{
An object of class \code{\link{bsseq}}
}
\description{
Convert \code{\link{methrix}} to \code{\link{bsseq}} object
}
\details{
Takes \code{\link{methrix}} object and returns a \code{\link{bsseq}} object
}
\examples{
\dontrun{
data("methrix_data")
methrix2bsseq(m = methrix_data)
}
}
