% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{methrix2bsseq}
\alias{methrix2bsseq}
\title{Convert methrix to bsseq object}
\usage{
methrix2bsseq(m)
}
\arguments{
\item{m}{\code{\link{methrix}} object}
}
\value{
An object of class \code{\link{BSseq}}
}
\description{
Convert methrix to bsseq object
}
\details{
Takes \code{\link{methrix}} object and returns a \code{BSseq} object
}
\examples{
\dontrun{
data("methrix_data")
methrix2bsseq(m = methrix_data)
}
}
