% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix-plot.R
\name{methrix_coverage}
\alias{methrix_coverage}
\title{Covergae QC Plots}
\usage{
methrix_coverage(m, type = c("hist", "dens"), pheno = NULL,
  perSample = F, return_ggplot = T)
}
\arguments{
\item{m}{Input \code{\link{methrix}} object}

\item{type}{Choose between "hist" (histogram) or "dens" (density plot).}

\item{pheno}{Column name of colData(m). Will be used as a factor to color different groups in the violin plot.}

\item{perSample}{Color the plots in a sample-wise manner?}

\item{return_ggplot}{Return a \code{\link{ggplot2}} object.}
}
\description{
Covergae QC Plots
}
