% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix-plot.R
\name{methrix_density}
\alias{methrix_density}
\title{Density Plot of \eqn{\beta}-Values}
\usage{
methrix_density(m, ranges = NULL, n_cpgs = 25000, pheno = NULL,
  bw.adjust = 2, col_palette = "RdYlGn")
}
\arguments{
\item{m}{Input \code{\link{methrix}} object}

\item{ranges}{genomic regions to be summarized. Could be a data.table with 3 columns (chr, start, end) or a \code{\link{GRanges}} object}

\item{n_cpgs}{Use these many random CpGs for plotting. Default 25000. Set it to \code{NULL} to use all - which can be memory expensive.}

\item{pheno}{Column name of colData(m). Will be used as a factor to color different groups in the violin plot.}

\item{bw.adjust}{Multiplicate bandwide adjustment. See \code{\link{geom_density}} for more information}

\item{col_palette}{Name of the ggplot diverging to use. Possible values: BrBG, PiYG, PRGn, PuOr, RdBu, RdGy, RdYlBu, RdYlGn, Spectral}
}
\value{
ggplot2 object
}
\description{
Density Plot of \eqn{\beta}-Values
}
\examples{
data("methrix_data")
methrix_density(m = methrix_data)
}
