% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix-plot.R
\name{methrix_pca}
\alias{methrix_pca}
\title{Principal Component Analysis}
\usage{
methrix_pca(m, ranges = NULL, pheno = NULL, do.plot = T, n_pc = 5)
}
\arguments{
\item{m}{Input \code{\link{methrix}} object}

\item{ranges}{\code{\link{GRanges}} object with GenomicRegions which should be used for subsetting.}

\item{pheno}{Column name of colData(m). Will be used as a factor to color different groups in the violin plot.}

\item{n_pc}{Number of principal components to return.}

\item{do.lot}{Should a plot be generated?}
}
\description{
Principal Component Analysis
}
