% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix-plot.R
\name{methrix_violin}
\alias{methrix_violin}
\title{Violin Plot for \eqn{\beta}-Values}
\usage{
methrix_violin(m, ranges = NULL, pheno = NULL, return_ggplot = FALSE)
}
\arguments{
\item{m}{Input \code{\link{methrix}} object}

\item{ranges}{genomic regions to be summarized. Could be a data.table with 3 columns (chr, start, end) or a \code{\link{GRanges}} object}

\item{pheno}{Column name of colData(m). Will be used as a factor to color different groups in the violin plot.}
}
\value{
ggplot2 object
}
\description{
Violin Plot for \eqn{\beta}-Values
}
\examples{
data("mm9_bsmap")
methrix_violin(m = mm9_bsmap)
}
