% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{order_by_sd}
\alias{order_by_sd}
\title{Order mathrix object by SD}
\usage{
order_by_sd(m)
}
\arguments{
\item{m}{\code{\link{methrix}} object}
}
\value{
An object of class \code{\link{methrix}}
}
\description{
Order mathrix object by SD
}
\details{
Takes \code{\link{methrix}} object and reorganizes the data by standard deviation
}
\examples{
data("methrix_data")
order_by_sd(m = methrix_data)
}
