% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bedgraphs.R
\name{read_bedgraphs}
\alias{read_bedgraphs}
\title{Versatile BedGraph reader.}
\usage{
read_bedgraphs(files = NULL, pipeline = NULL, ideal = FALSE,
  vect = TRUE, vect_batch_size = NULL, coldata = NULL,
  chr_idx = NULL, start_idx = NULL, end_idx = NULL,
  beta_idx = NULL, M_idx = NULL, U_idx = NULL, strand_idx = NULL,
  cov_idx = NULL, h5 = FALSE, h5_dir = NULL, verbose = TRUE)
}
\arguments{
\item{files}{bedgraph files.}

\item{pipeline}{Default NULL. Can be "Bismark" or "MethylDeckal". If not known use idx arguments for manual column assignments.}

\item{ideal}{If all bedgraphs contain same number of CpG's and in same order. If TRUE instead of merging tables, matrices are crated by simply cbinding which is significantly faster.}

\item{vect}{To use vectorized code. Default TRUE, memory intese. Set to FALSE if you have large number of BedGraph files.}

\item{vect_batch_size}{Default NULL. Process samples in batches. Applicable only when vect = TRUE}

\item{coldata}{An optional DataFrame describing the samples. Row names, if present, become the column names of the matrix. If NULL, then a DataFrame will be created with basename of files used as the row names.}

\item{chr_idx}{column index for chromosome in bedgraph files}

\item{start_idx}{column index for start position in bedgraph files}

\item{end_idx}{column index for end position in bedgraph files}

\item{beta_idx}{column index for beta values in bedgraph files}

\item{M_idx}{column index for read counts supporting Methylation in bedgraph files}

\item{U_idx}{column index for read counts supporting Un-methylation in bedgraph files}

\item{strand_idx}{column index for strand information in bedgraph files}

\item{cov_idx}{column index for total-coverage in bedgraph files}

\item{h5}{Should the coverage and methylation matrices be stored as "HDF5Array"}

\item{h5_dir}{directory to store H5 based object}

\item{verbose}{Be little chatty ? Default TRUE.}
}
\description{
Versatile BedGraph reader.
}
\details{
Reads BedGraph files and generates methylation and coverage matrices. Optionally arrays can be serialized as on-disk HDFS5 arrays.
}
