% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methrix_operations.R
\name{region_filter}
\alias{region_filter}
\title{Filter matrices by region}
\usage{
region_filter(m, regions)
}
\arguments{
\item{m}{\code{\link{methrix}} object}

\item{regions}{genomic regions to subset by. Could be a data.table with 3 columns (chr, start, end) or a \code{\link{GRanges}} object}
}
\description{
Filter matrices by region
}
\details{
Takes \code{\link{methrix}} object and filters CpGs based on supplied regions in data.table or GRanges format
}
