% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bedgraphs.R
\name{write_bedgraphs}
\alias{write_bedgraphs}
\title{Writes bedGraphs from methrix object}
\usage{
write_bedgraphs(m, output_dir = NULL, rm_NA = TRUE, force = FALSE,
  n_thr = 4, make_BigWig = FALSE, bedtools_path = NULL,
  bedGraphToBigWig_path = NULL, bedClip_path = NULL)
}
\arguments{
\item{m}{\code{\link{methrix}} object}

\item{output_dir}{Output directory name where the files should be saved. If \code{NULL} creats a \code{tempdir}}

\item{rm_NA}{remove NAs}

\item{force}{forces to create files if they are existing}

\item{n_thr}{Default 4.}

\item{bedtools_path}{Default NULL. Path to bedtools. Default looks in $PATH. Only applicable if \code{make_bigWig} = TRUE}

\item{bedGraphToBigWig_path}{Default NULL. Path to bedGraphToBigWig Default looks in $PATH. Only applicable if \code{make_bigWig} = TRUE}

\item{bedClip_path}{Default NULL. Path to bedClip Default looks in $PATH. Only applicable if \code{make_bigWig} = TRUE}

\item{make_bigWig}{Default FALSE. Whether to convert bedGraph to BigWig}
}
\value{
writes bedgraph files to output
}
\description{
Writes bedGraphs from methrix object
}
\examples{
data("mm9_bsmap")
write_bedgraphs(m = mm9_bsmap, output_dir = "./temp")
}
