% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bedgraphs.R
\name{write_bedgraphs}
\alias{write_bedgraphs}
\title{Writes bedGraphs from methrix object}
\usage{
write_bedgraphs(
  m,
  output_dir = NULL,
  rm_NA = TRUE,
  force = FALSE,
  n_thr = 4,
  compress = TRUE
)
}
\arguments{
\item{m}{\code{\link{methrix}} object}

\item{output_dir}{Output directory name where the files should be saved.
If \code{NULL} creats a \code{tempdir}}

\item{rm_NA}{remove NAs}

\item{force}{forces to create files if they are existing}

\item{n_thr}{Default 4.}

\item{compress}{Whether to compress the output. Default TRUE}
}
\value{
writes bedgraph files to output
}
\description{
Writes bedGraphs from methrix object
}
\examples{
data('methrix_data')
write_bedgraphs(m = methrix_data, output_dir = './temp')
}
